src/test/javascript/spec/app/entities/entidades/entidades-update.component.spec.ts
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder } from '@angular/forms';
import { of } from 'rxjs';

import { GestionEmpleadosTestModule } from '../../../test.module';
import { EntidadesUpdateComponent } from 'app/entities/entidades/entidades-update.component';
import { EntidadesService } from 'app/entities/entidades/entidades.service';
import { Entidades } from 'app/shared/model/entidades.model';

describe('Component Tests', () => {
  describe('Entidades Management Update Component', () => {
    let comp: EntidadesUpdateComponent;
    let fixture: ComponentFixture<EntidadesUpdateComponent>;
    let service: EntidadesService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [GestionEmpleadosTestModule],
        declarations: [EntidadesUpdateComponent],
        providers: [FormBuilder],
      })
        .overrideTemplate(EntidadesUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(EntidadesUpdateComponent);
      comp = fixture.componentInstance;
      service = fixture.debugElement.injector.get(EntidadesService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new Entidades(123);
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new Entidades();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
